@@ -1,5 +1,6 @@
 {
   "name": "chatgpt-web",
+  "type": "module",
   "version": "2.13.3",
   "private": false,
   "description": "ChatGPT Web",
@@ -26,16 +27,16 @@
     "@antfu/eslint-config": "^0.43.1",
     "@commitlint/cli": "^17.4.4",
     "@commitlint/config-conventional": "^17.4.4",
-    "@iconify/vue": "^4.1.0",
+    "@iconify/vue": "^4.1.2",
     "@traptitech/markdown-it-katex": "^3.6.0",
     "@types/chart.js": "^2.9.37",
     "@types/katex": "^0.16.0",
     "@types/markdown-it": "^12.2.3",
     "@types/markdown-it-link-attributes": "^3.0.1",
     "@types/node": "^18.14.6",
-    "@vitejs/plugin-vue": "^5.0.4",
-    "@vueuse/core": "^10.9.0",
-    "autoprefixer": "^10.4.13",
+    "@vitejs/plugin-vue": "^5.1.4",
+    "@vueuse/core": "^10.11.1",
+    "autoprefixer": "^10.4.20",
     "axios": "^1.6.7",
     "chart.js": "^4.3.0",
     "dayjs": "^1.11.7",
@@ -49,21 +50,21 @@
     "lint-staged": "^13.1.2",
     "markdown-it": "^13.0.1",
     "markdown-it-link-attributes": "^4.0.1",
-    "naive-ui": "^2.38.2",
+    "naive-ui": "^2.40.1",
     "npm-run-all": "^4.1.5",
-    "pinia": "^2.0.33",
-    "postcss": "^8.4.35",
+    "pinia": "^2.2.4",
+    "postcss": "^8.4.47",
     "qrcode.vue": "^3.4.1",
     "rimraf": "^4.3.0",
-    "tailwindcss": "^3.4.1",
+    "tailwindcss": "^3.4.14",
     "typescript": "5.3.3",
-    "vite": "^5.1.7",
-    "vite-plugin-pwa": "^0.20.0",
-    "vue": "^3.4.21",
-    "vue-chartjs": "^5.2.0",
-    "vue-i18n": "^9.10.1",
-    "vue-router": "^4.3.0",
-    "vue-tsc": "^2.0.6"
+    "vite": "^5.4.10",
+    "vite-plugin-pwa": "^0.20.5",
+    "vue": "^3.5.12",
+    "vue-chartjs": "^5.3.1",
+    "vue-i18n": "9.13.1",
+    "vue-router": "^4.4.5",
+    "vue-tsc": "^2.1.6"
   },
   "lint-staged": {
     "*.{ts,tsx,vue}": [
